etl/etl/requirements.txt
adlfs==2024.7.0 ; python_version >= "3.10" and python_version < "4.0"
aiohappyeyeballs==2.4.3 ; python_version >= "3.10" and python_version < "4.0"
aiohttp==3.11.1 ; python_version >= "3.10" and python_version < "4.0"
aiosignal==1.3.1 ; python_version >= "3.10" and python_version < "4.0"
asciitree==0.3.3 ; python_version >= "3.10" and python_version < "4.0"
async-timeout==5.0.1 ; python_version >= "3.10" and python_version < "3.11"
attrs==24.2.0 ; python_version >= "3.10" and python_version < "4.0"
azure-core==1.32.0 ; python_version >= "3.10" and python_version < "4.0"
azure-datalake-store==0.0.53 ; python_version >= "3.10" and python_version < "4.0"
azure-functions==1.21.3 ; python_version >= "3.10" and python_version < "4.0"
azure-identity==1.19.0 ; python_version >= "3.10" and python_version < "4.0"
azure-monitor-opentelemetry-exporter==1.0.0b32 ; python_version >= "3.10" and python_version < "4.0"
azure-storage-blob==12.24.0 ; python_version >= "3.10" and python_version < "4.0"
azure-storage-file-datalake==12.18.0 ; python_version >= "3.10" and python_version < "4.0"
certifi==2024.8.30 ; python_version >= "3.10" and python_version < "4.0"
cffi==1.17.1 ; python_version >= "3.10" and python_version < "4.0"
cfgrib==0.9.14.1 ; python_version >= "3.10" and python_version < "4.0"
cftime==1.6.4.post1 ; python_version >= "3.10" and python_version < "4.0"
charset-normalizer==3.4.0 ; python_version >= "3.10" and python_version < "4.0"
click==8.1.7 ; python_version >= "3.10" and python_version < "4.0"
colorama==0.4.6 ; python_version >= "3.10" and python_version < "4.0" and platform_system == "Windows"
cryptography==43.0.3 ; python_version >= "3.10" and python_version < "4.0"
deprecated==1.2.15 ; python_version >= "3.10" and python_version < "4.0"
eccodes==2.38.3 ; python_version >= "3.10" and python_version < "4.0"
fasteners==0.19 ; python_version >= "3.10" and python_version < "4.0" and sys_platform != "emscripten"
findlibs==0.0.5 ; python_version >= "3.10" and python_version < "4.0"
fixedint==0.1.6 ; python_version >= "3.10" and python_version < "4.0"
frozenlist==1.5.0 ; python_version >= "3.10" and python_version < "4.0"
fsspec==2024.10.0 ; python_version >= "3.10" and python_version < "4.0"
idna==3.10 ; python_version >= "3.10" and python_version < "4.0"
importlib-metadata==8.5.0 ; python_version >= "3.10" and python_version < "4.0"
isodate==0.7.2 ; python_version >= "3.10" and python_version < "4.0"
msal-extensions==1.2.0 ; python_version >= "3.10" and python_version < "4.0"
msal==1.31.0 ; python_version >= "3.10" and python_version < "4.0"
msrest==0.7.1 ; python_version >= "3.10" and python_version < "4.0"
multidict==6.1.0 ; python_version >= "3.10" and python_version < "4.0"
numcodecs==0.13.1 ; python_version >= "3.10" and python_version < "4.0"
numpy==2.1.3 ; python_version >= "3.10" and python_version < "4.0"
oauthlib==3.2.2 ; python_version >= "3.10" and python_version < "4.0"
opentelemetry-api==1.28.2 ; python_version >= "3.10" and python_version < "4.0"
opentelemetry-sdk==1.28.2 ; python_version >= "3.10" and python_version < "4.0"
opentelemetry-semantic-conventions==0.49b2 ; python_version >= "3.10" and python_version < "4.0"
packaging==24.2 ; python_version >= "3.10" and python_version < "4.0"
pandas==2.2.3 ; python_version >= "3.10" and python_version < "4.0"
portalocker==2.10.1 ; python_version >= "3.10" and python_version < "4.0"
propcache==0.2.0 ; python_version >= "3.10" and python_version < "4.0"
psutil==5.9.8 ; python_version >= "3.10" and python_version < "4.0"
pycparser==2.22 ; python_version >= "3.10" and python_version < "4.0"
pyjwt[crypto]==2.9.0 ; python_version >= "3.10" and python_version < "4.0"
python-dateutil==2.9.0.post0 ; python_version >= "3.10" and python_version < "4.0"
python-dotenv==1.0.1 ; python_version >= "3.10" and python_version < "4.0"
pytz==2024.2 ; python_version >= "3.10" and python_version < "4.0"
pywin32==308 ; python_version >= "3.10" and python_version < "4.0" and platform_system == "Windows"
requests-oauthlib==2.0.0 ; python_version >= "3.10" and python_version < "4.0"
requests==2.32.3 ; python_version >= "3.10" and python_version < "4.0"
scipy==1.14.1 ; python_version >= "3.10" and python_version < "4.0"
six==1.16.0 ; python_version >= "3.10" and python_version < "4.0"
typing-extensions==4.12.2 ; python_version >= "3.10" and python_version < "4.0"
tzdata==2024.2 ; python_version >= "3.10" and python_version < "4.0"
urllib3==2.2.3 ; python_version >= "3.10" and python_version < "4.0"
wrapt==1.16.0 ; python_version >= "3.10" and python_version < "4.0"
xarray==2024.10.0 ; python_version >= "3.10" and python_version < "4.0"
yarl==1.17.1 ; python_version >= "3.10" and python_version < "4.0"
zarr==2.18.3 ; python_version >= "3.10" and python_version < "4.0"
zipp==3.21.0 ; python_version >= "3.10" and python_version < "4.0"
